[#639] SpeechService·PlaceService 신설 — 음성·지도 SDK 구현체 분리
— Speech·AVFoundation 결합 2파일, MapKit 기반 PlaceSuggestEngine 구현체 이동. PlaceSuggestUsecaseImple은 SDK 무의존이라 Domain 잔류
— 테스트 없는 프레임워크라 framework() 사용 — 테스트 스킴 미생성
