+++ Services/PlaceService/PlaceService.h
@@ -1,0 +1,17 @@
+//
+//  PlaceService.h
+//  PlaceService
+//
+//  Created by sudo.park on 7/30/26.
+//
+
+#import <Foundation/Foundation.h>
+
+//! Project version number for PlaceService.
+FOUNDATION_EXPORT double PlaceServiceVersionNumber;
+
+//! Project version string for PlaceService.
+FOUNDATION_EXPORT const unsigned char PlaceServiceVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <PlaceService/PublicHeader.h>
+
