+++ Services/SpeechService/Project.swift
@@ -1,0 +1,22 @@
+import ProjectDescriptionHelpers
+import ProjectDescription
+
+let project = Project.framework(
+    name: "SpeechService",
+    destinations: [.iPhone],
+    iOSTargetVersion: "17.0",
+    dependencies: [
+        .project(
+            target: "Common3rdParty",
+            path: .relativeToCurrentFile("../../Supports/Common3rdParty")
+        ),
+        .project(
+            target: "Domain",
+            path: .relativeToCurrentFile("../../Domain")
+        ),
+        .project(
+            target: "Extensions",
+            path: .relativeToCurrentFile("../../Supports/Extensions")
+        )
+    ]
+)
