+++ Services/SpeechService/Sources/SpeechRecognizePermissionCheckerImple.swift
@@ -1,6 +1,6 @@
 //
 //  SpeechRecognizePermissionCheckerImple.swift
-//  Domain
+//  SpeechService
 //
 //  Created by sudo.park on 6/7/26.
 //  Copyright © 2026 com.sudo.park. All rights reserved.
@@ -9,6 +9,7 @@
 import Foundation
 import Speech
 import AVFoundation
+import Domain
 
 
 public final class SpeechRecognizePermissionCheckerImple: SpeechRecognizePermissionChecker, @unchecked Sendable {
