+++ Services/SpeechService/SpeechService.h
@@ -1,0 +1,17 @@
+//
+//  SpeechService.h
+//  SpeechService
+//
+//  Created by sudo.park on 7/30/26.
+//
+
+#import <Foundation/Foundation.h>
+
+//! Project version number for SpeechService.
+FOUNDATION_EXPORT double SpeechServiceVersionNumber;
+
+//! Project version string for SpeechService.
+FOUNDATION_EXPORT const unsigned char SpeechServiceVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <SpeechService/PublicHeader.h>
+
