+++ Workspace.swift
@@ -21,6 +21,8 @@ let workspace = Workspace(name: "TodoCalendar",
                             "Presentations/EventListScenes",
                             "Presentations/AIAgentScene",
                             "Services/FirstPartyServices",
+                            "Services/SpeechService",
+                            "Services/PlaceService",
                             "Supports/Extensions",
                             "Supports/UnitTestHelpKit",
                             "Supports/Common3rdParty",
